Revert "Merge pull request #84 from caiotheodoro/fix/judge-superset"
This reverts commit 533a6634c2cad4112bc42f869903b7ef4b39cb42, reversing
changes made to e7a735ccf284d993f7f7242ca5519989425596a7.
