+++ packages/eval/src/llm/judge-config.ts
@@ -79,7 +79,6 @@ function resolveEntry(
     apiKey: entry.apiKeyEnv ? process.env[entry.apiKeyEnv] : envDefaults.apiKey,
     temperature: entry.temperature ?? envDefaults.temperature,
     timeoutMs: entry.timeoutMs ?? envDefaults.timeoutMs,
-    retryCount: entry.retryCount,
   };
 }
