+++ packages/eval/src/test/judge-json-salvage.test.ts
@@ -69,7 +69,6 @@ describe("llmJudge parse-failure defenses", () => {
         return Promise.resolve({
           ok: true,
           status: 200,
-          headers: new Map(),
           json: async () => ({
             choices: [{ message: { content: typeof content === "string" ? content : "" } }],
           }),
