+++ packages/shared/package.json
@@ -18,7 +18,6 @@
     "typecheck": "tsc -p tsconfig.json --noEmit"
   },
   "devDependencies": {
-    "@types/node": "^26.2.0",
     "typescript": "^5.5.0",
     "vitest": "^2.0.0"
   },
